Seperate Component done
