+++ src/pages/Register/SignupData.scss
@@ -1,0 +1,1 @@
+@import '/src/styles/variables.scss';
